Corrected initialization coordinates

+++ simulator/backend/Particle.h
@@ -16,7 +16,7 @@ struct Point {
 
 class Particle {
  public:
-  Particle(int x, int y, SimulatorSettings* settings) : x_(x), y_(y), settings_(settings) {}
+  Particle(double x, double y, SimulatorSettings* settings) : x_(x), y_(y), settings_(settings) {}
   ~Particle();
 
   Point point() const { return Point(x_, y_); }
